formatting cleanup for DataLexicaliser
